Endpoint information from ToxCast

Downloaded on October 2015 from ToxCast. The file name of the raw data was "Assay_Summary_151020.csv" from the zip file "Assay_Information_Oct_2015.zip". At the time of the toxEval package release, these data were found at: https://www.epa.gov/chemical-research/toxicity-forecaster-toxcasttm-data in the section marked "Download Assay Information", in the ToxCast & Tox21 high-throughput assay information data set.

Source

https://www.epa.gov/chemical-research/toxcast-assay-annotation-data-user-guide

Value

data frame with 86 columns. The columns and definitions are discussed in the "ToxCast Assay Annotation Version 1.0 Data User Guide (PDF)" (see source)

References

U.S. EPA. 2014. ToxCast Assay Annotation Data User Guide. Retrieved from https://www.epa.gov/chemical-research/toxcast-assay-annotation-data-user-guide.

Examples

end_point_info <- end_point_info
head(end_point_info[,1:5])

#> # A tibble: 6 x 5
#>     aid  acid  aeid  asid assay_source_name
#>   <int> <int> <int> <int> <chr>            
#> 1     1     1     1     1 ACEA             
#> 2     1     1     2     1 ACEA             
#> 3     2     2     3     2 APR              
#> 4     2     2     4     2 APR              
#> 5     2     3     5     2 APR              
#> 6     2     3     6     2 APR