Get the ACC values for a selection of chemicals

The get_ACC function retrieves the activity concentration at cutoff (ACC) values for specified chemicals.

Usage

get_ACC(CAS)

Arguments

CAS

Vector of CAS.

Value

data frame with columns CAS, chnm, flags, endPoint, ACC, MlWt, and ACC_value

Details

The data used in toxEval were combined from files in the "INVITRODB_V3_LEVEL5" directory that were included in the October 2018 release of the ToxCast database. The function get_ACC will convert the ACC values in the ToxCast database from units of (log \(\mu\)M) to units of \(\mu\)g/L, and reformat the data as input to toxEval.

Examples

CAS <- c("121-00-6", "136-85-6", "80-05-7", "84-65-1", "5436-43-1", "126-73-8")
ACC <- get_ACC(CAS)
head(ACC)
#>        CAS                                                           flags
#> 1  80-05-7                                                            <NA>
#> 2 126-73-8 Only highest conc above baseline, active|Less than 50% efficacy
#> 3 121-00-6                        Only highest conc above baseline, active
#> 4  80-05-7                                                            <NA>
#> 5 121-00-6                        Only highest conc above baseline, active
#> 6  80-05-7                                                            <NA>
#>                           endPoint      ACC    MlWt ACC_value
#> 1    ACEA_AR_agonist_AUC_viability 1.137013 228.291  3129.692
#> 2    ACEA_AR_agonist_AUC_viability 1.796335 266.318 16662.331
#> 3    ACEA_AR_agonist_AUC_viability 1.575112 180.247  6776.111
#> 4          ACEA_AR_antagonist_80hr 1.065844 228.291  2656.643
#> 5          ACEA_AR_antagonist_80hr 1.559357 180.247  6534.693
#> 6 ACEA_AR_antagonist_AUC_viability 1.095331 228.291  2843.279
#>                           chnm
#> 1                  Bisphenol A
#> 2           Tributyl phosphate
#> 3 2-tert-Butyl-4-methoxyphenol
#> 4                  Bisphenol A
#> 5 2-tert-Butyl-4-methoxyphenol
#> 6                  Bisphenol A