Compute EAR values

This function computes Exposure:Activity ratios using user-provided measured concentration data from the output of create_toxEval, and joins the data with the activity concentration at cutoff data provided by ToxCast.Data from ToxCast is included with this package, but alternative benchmark data can be provided to perform the same "toxEval" analysis.

get_chemical_summary(
  tox_list,
  ACC = NULL,
  filtered_ep = "All",
  chem_data = NULL,
  chem_site = NULL,
  chem_info = NULL,
  exclusion = NULL
)

Arguments

tox_list
List with data frames for chem_data, chem_info, chem_site, and optionally exclusions and benchmarks. Created with create_toxEval.

ACC
Data frame with columns: CAS, chnm, endPoint, and ACC_value for specific chemical/endpoint combinations generated using the get_ACC function. EndPoints with specific data quality flags may optionally be removed using the remove_flags function.

filtered_ep
Data frame with columns: endPoints, groupCol. Default is "All", where no filtering occurs.

chem_data
Optional data frame with (at least) columns: CAS, SiteID, and Value. Default is NULL. The argument will over-ride what is in tox_list.

chem_site
Optional data frame with (at least) columns: SiteID, and Short Name. Default is NULL. The argument will over-ride what is in tox_list.

chem_info
Optional data frame with (at least) columns: CAS, and class. Default is NULL. The argument will over-ride what is in tox_list.

exclusion
Optional data frame with (at least) columns: CAS and endPoint. Default is NULL. The argument will over-ride what is in tox_list.

Value

a data frame with the columns: CAS, chnm (chemical name as a factor), site, date, EAR, Bio_category, shortName (of site), Class. The output of this function is where you find EAR values for every chemical/endpoint combination.

Details

To use the data provided by the package, a sample workflow is shown below in the examples. The examples include retrieving the ToxCast (ACC) values that are used to calculate EARs, choosing endPoints that should be ignored based on data quality "flags" in the ToxCast database, and removing groups of endPoints that may not be important to the analysis at hand.

Examples

path_to_tox <-  system.file("extdata", package="toxEval")
file_name <- "OWC_data_fromSup.xlsx"
full_path <- file.path(path_to_tox, file_name)

tox_list <- create_toxEval(full_path)

ACC <- get_ACC(tox_list$chem_info$CAS)
ACC <- remove_flags(ACC)

cleaned_ep <- clean_endPoint_info(end_point_info)
filtered_ep <- filter_groups(cleaned_ep)

chemical_summary <- get_chemical_summary(tox_list, ACC, filtered_ep)
head(chemical_summary)

#>       CAS          chnm                  endPoint          site
#> 1 84-65-1 Anthraquinone TOX21_ERa_LUC_VM7_Agonist USGS-04024000
#> 2 84-65-1 Anthraquinone TOX21_ERa_LUC_VM7_Agonist USGS-04024000
#> 3 84-65-1 Anthraquinone TOX21_ERa_LUC_VM7_Agonist USGS-04067500
#> 4 84-65-1 Anthraquinone TOX21_ERa_LUC_VM7_Agonist USGS-04067500
#> 5 84-65-1 Anthraquinone TOX21_ERa_LUC_VM7_Agonist USGS-04067500
#> 6 84-65-1 Anthraquinone TOX21_ERa_LUC_VM7_Agonist USGS-04067500
#>                  date          EAR     Bio_category shortName         Class
#> 1 2011-05-01 13:00:00 1.804762e-05 Nuclear Receptor   StLouis Dyes/Pigments
#> 2 2013-04-27 10:40:00 9.023810e-06 Nuclear Receptor   StLouis Dyes/Pigments
#> 3 2011-03-21 15:36:00 1.503968e-05 Nuclear Receptor Menominee Dyes/Pigments
#> 4 2011-04-26 12:15:00 9.023810e-06 Nuclear Receptor Menominee Dyes/Pigments
#> 5 2011-04-28 13:48:00 6.015873e-06 Nuclear Receptor Menominee Dyes/Pigments
#> 6 2011-06-21 09:56:00 9.023810e-06 Nuclear Receptor Menominee Dyes/Pigments