Biological hits per category

The hits_by_groupings_DT (DT option) and hits_by_groupings (data frame option) functions create tables with one row per category("Biological", "Chemical", or "Chemical Class"). The columns indicate the "Biological" groupings. The values in the table signify how many sites have samples with EARs that exceeded the hit_threshold for that particular "Biological"/category combination. If the user chooses "Biological" as the category, it is a simple 2-column table of "Biological" groupings and number of sites (nSites).

Usage

hits_by_groupings_DT(
  chemical_summary,
  category = "Biological",
  mean_logic = FALSE,
  sum_logic = TRUE,
  hit_threshold = 0.1
)

hits_by_groupings(
  chemical_summary,
  category,
  mean_logic = FALSE,
  sum_logic = TRUE,
  hit_threshold = 0.1
)

Arguments

chemical_summary

Data frame from get_chemical_summary.

category

Character. Either "Biological", "Chemical Class", or "Chemical".

mean_logic

Logical. TRUE displays the mean sample from each site, FALSE displays the maximum sample from each site.

sum_logic

Logical. TRUE sums the EARs in a specified grouping, FALSE does not. FALSE may be better for traditional benchmarks as opposed to ToxCast benchmarks.

hit_threshold

Numeric threshold defining a "hit".

Value

data frame with one row per category, and one column per Biological grouping.

Details

The tables result in slightly different results for a single site, displaying the number of samples with hits rather than the number of sites.

Examples

# This is the example workflow:
path_to_tox <- system.file("extdata", package = "toxEval")
file_name <- "OWC_data_fromSup.xlsx"

full_path <- file.path(path_to_tox, file_name)

tox_list <- create_toxEval(full_path)

ACC <- get_ACC(tox_list$chem_info$CAS)
ACC <- remove_flags(ACC)

cleaned_ep <- clean_endPoint_info(end_point_info)
filtered_ep <- filter_groups(cleaned_ep)
chemical_summary <- get_chemical_summary(tox_list, ACC, filtered_ep)

site_df <- hits_by_groupings(chemical_summary, category = "Biological")

hits_by_groupings_DT(chemical_summary, category = "Biological")

hits_by_groupings_DT(chemical_summary, category = "Chemical Class")

hits_by_groupings_DT(chemical_summary, category = "Chemical")