Filter endPoints based on groups and assays.

This function provides a mechanism to specify 3 levels of information in the supplied data frame end_point_info to be used in subsequent analysis steps. First, the user specifies the ToxCast assay annotation using the 'groupCol' argument, which is a column header in 'end_point_info'. Second, the user specifies the families of assays to use. Finally, the user can choose to remove specific group(s) from the category. The default is to remove 'Background Measurement' and 'Undefined'. Choices for this should be reconsidered based on individual study objectives.

Usage

filter_groups(
  ep,
  groupCol = "intended_target_family",
  assays = c("ACEA", "APR", "ATG", "NVS", "OT", "TOX21", "CEETOX", "LTEA", "CLD",
    "TANGUAY", "CCTE_PADILLA", "CCTE", "STM", "ARUNA", "CCTE_SHAFER", "CPHEA_STOKER",
    "CCTE_GLTED", "UPITT", "UKN", "ERF", "TAMU", "IUF", "CCTE_MUNDY", "UTOR", "VALA"),
  remove_groups = c("Background Measurement", "Undefined")
)

Arguments

ep

Data frame containing Endpoint information from ToxCast

groupCol

Character name of ToxCast annotation column to use as a group category

assays

Vector of assays to use in the data analysis. Possible values are "ACEA", "APR", "ATG", "NVS", "OT", "TOX21", "CEETOX", "LTEA", "CLD", "TANGUAY", "CCTE_PADILLA", "BSK" , "CCTE", "STM", "ARUNA", "CCTE_SHAFER", "CPHEA_STOKER", "CCTE_GLTED", "UPITT", "UKN", "ERF", "TAMU", "IUF", "CCTE_MUNDY", "UTOR", "VALA". By default, the "BSK" (BioSeek) assay is removed.

remove_groups

Vector of groups within the selected 'groupCol' to remove.

Details

The default category ('groupCol') is 'intended_target_family'. Depending on the study, other categories may be more relevant. The best resource on these groupings is the "ToxCast Assay Annotation Data User Guide" directly from EPA https://www.epa.gov/chemical-research/toxcast-assay-annotation-data-user-guide. Following that link, it defines "intended_target_family" as "the target family of the objective target for the assay". Much more detail can be discovered in that documentation.

Examples

end_point_info <- end_point_info
cleaned_ep <- clean_endPoint_info(end_point_info)
filtered_ep <- filter_groups(cleaned_ep)
head(filtered_ep)
#>                          endPoint         groupCol assaysFull
#> 1                    ACEA_ER_80hr Nuclear Receptor       ACEA
#> 2           ACEA_ER_AUC_viability       Cell Cycle       ACEA
#> 3 APR_HepG2_CellCycleArrest_1h_dn       Cell Cycle        APR
#> 4 APR_HepG2_CellCycleArrest_1h_up       Cell Cycle        APR
#> 5        APR_HepG2_CellLoss_1h_dn       Cell Cycle        APR
#> 6        APR_HepG2_CellLoss_1h_up       Cell Cycle        APR