Compute EAR values

This function computes Exposure:Activity ratios using user-provided measured concentration data from the output of create_toxEval, and joins the data with the activity concentration at cutoff data provided by ToxCast.Data from ToxCast is included with this package, but alternative benchmark data can be provided to perform the same "toxEval" analysis.

Usage

get_chemical_summary(
  tox_list,
  ACC = NULL,
  filtered_ep = "All",
  chem_data = NULL,
  chem_site = NULL,
  chem_info = NULL,
  exclusion = NULL
)

Arguments

tox_list

List with data frames for chem_data, chem_info, chem_site, and optionally exclusions and benchmarks. Created with create_toxEval.

ACC

Data frame with columns: CAS, chnm, endPoint, and ACC_value for specific chemical/endpoint combinations generated using the get_ACC function. EndPoints with specific data quality flags may optionally be removed using the remove_flags function.

filtered_ep

Data frame with columns: endPoints, groupCol. Default is "All", where no filtering occurs.

chem_data

Optional data frame with (at least) columns: CAS, SiteID, and Value. Default is NULL. The argument will over-ride what is in tox_list.

chem_site

Optional data frame with (at least) columns: SiteID, and Short Name. Default is NULL. The argument will over-ride what is in tox_list.

chem_info

Optional data frame with (at least) columns: CAS, and class. Default is NULL. The argument will over-ride what is in tox_list.

exclusion

Optional data frame with (at least) columns: CAS and endPoint. Default is NULL. The argument will over-ride what is in tox_list.

Value

a data frame with the columns: CAS, chnm (chemical name as a factor), site, date, EAR, Bio_category, shortName (of site), Class. The output of this function is where you find EAR values for every chemical/endpoint combination.

Details

To use the data provided by the package, a sample workflow is shown below in the examples. The examples include retrieving the ToxCast (ACC) values that are used to calculate EARs, choosing endPoints that should be ignored based on data quality "flags" in the ToxCast database, and removing groups of endPoints that may not be important to the analysis at hand.

Examples

path_to_tox <- system.file("extdata", package = "toxEval")
file_name <- "OWC_data_fromSup.xlsx"
full_path <- file.path(path_to_tox, file_name)

tox_list <- create_toxEval(full_path)

ACC <- get_ACC(tox_list$chem_info$CAS)
ACC <- remove_flags(ACC)

cleaned_ep <- clean_endPoint_info(end_point_info)
filtered_ep <- filter_groups(cleaned_ep)

chemical_summary <- get_chemical_summary(tox_list, ACC, filtered_ep)
head(chemical_summary)
#>        CAS              chnm             endPoint          site
#> 1 119-36-8 Methyl salicylate ACEA_AR_agonist_80hr USGS-04024000
#> 2 119-36-8 Methyl salicylate ACEA_AR_agonist_80hr USGS-04095090
#> 3 119-36-8 Methyl salicylate ACEA_AR_agonist_80hr USGS-04137500
#> 4 119-36-8 Methyl salicylate ACEA_AR_agonist_80hr USGS-04137500
#> 5 119-36-8 Methyl salicylate ACEA_AR_agonist_80hr USGS-04137500
#> 6 119-36-8 Methyl salicylate ACEA_AR_agonist_80hr USGS-04137500
#>                  date          EAR     Bio_category shortName         Class
#> 1 2011-05-01 13:00:00 2.223770e-05 Nuclear Receptor   StLouis Miscellaneous
#> 2 2011-10-20 13:45:00 1.779016e-05 Nuclear Receptor     Burns Miscellaneous
#> 3 2011-11-10 14:00:00 2.668524e-05 Nuclear Receptor   AuSable Miscellaneous
#> 4 2011-12-13 15:00:00 4.002786e-05 Nuclear Receptor   AuSable Miscellaneous
#> 5 2012-01-10 11:00:00 2.668524e-05 Nuclear Receptor   AuSable Miscellaneous
#> 6 2012-02-21 13:00:00 5.337048e-05 Nuclear Receptor   AuSable Miscellaneous