Remove endpoints with specific data quality flags from data

Through the ToxCast program quality assurance procedures, information is examined and at times, it is necessary to assign a data quality flag to a specific chemical/assay result. A toxEval user may want to include or exclude assay results with certain flags depending on the objectives of a given study. Assay results with specific data quality flags assigned to them can be removed based on their designated flag with the remove_flags function. The flags included in ToxCast, and the associated flagsShort value (used in the remove_flags function) are as follows:

Flag	flagsShort

Borderline active*	Borderline*

Only highest conc above baseline, active*	OnlyHighest*

Only one conc above baseline, active	OneAbove

Noisy data	Noisy

Hit-call potentially confounded by overfitting	HitCall

Gain AC50 < lowest conc & loss AC50 < mean conc*	GainAC50*

Biochemical assay with < 50% efficacy*	Biochemical*

Less than 50% efficacy	LessThan50

AC50 less than lowest concentration tested*	ACCLessThan*

GNLSmodel	GNLSmodel

Asterisks indicate flags removed in the function as default.

remove_flags(
  ACC,
  flagsShort = c("Borderline", "OnlyHighest", "GainAC50", "Biochemical", "ACCLessThan")
)

Arguments

ACC
data frame with columns: casn, chnm, endPoint, and ACC_value

flagsShort
vector of flags to to trigger REMOVAL of chemical:endPoint combination. Possible values are "Borderline", "OnlyHighest", "OneAbove", "Noisy", "HitCall", "GainAC50", "Biochemical","LessThan50","ACCLessThan","GNLSmodel".

Examples

CAS <- c("121-00-6","136-85-6","80-05-7","84-65-1","5436-43-1","126-73-8")
ACC <- get_ACC(CAS)
nrow(ACC)

#> [1] 362

ACC <- remove_flags(ACC)
nrow(ACC)

#> [1] 219